RR-173: Extracted event form component to solve Form specifics
